[Chore] Factor out utils into separate module
Problem: `wizard_structure` module contains too many definitions,
which became hard to maintain.

Solution: Factor out utils into separate module.
