+++ baking/src/tezos_baking/tezos_voting_wizard.py
@@ -12,7 +12,8 @@ import os, sys
 import readline
 import re
 
-from .wizard_structure import *
+from tezos_baking.wizard_structure import *
+from tezos_baking.util import *
 
 # Global options
